pyproject.toml
[tool.poetry]
name = "utter"
version = "0.1.0"
description = "backstage of the utter stories"
authors = ["Victor Oldensand <victormag24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
